rename : manageClub => clubManage
