+++ src/routes/router.tsx
@@ -9,9 +9,9 @@ import ProfilePage from '@/pages/ProfilePage/ProfilePage';
 import RegisterPage from '@/pages/RegisterPage/RegisterPage';
 import ClubEventPage from '@/pages/club/ClubEventPage/ClubEventPage';
 import ClubHomePage from '@/pages/club/ClubHomePage/ClubHomePage';
+import ClubManagePage from '@/pages/club/ClubManagePage/ClubManagePage';
 import CreateClubPage from '@/pages/club/CreateClubPage/CreateClubPage';
 import InvitePage from '@/pages/club/InvitePage/InvitePage';
-import ManageClubPage from '@/pages/club/ManageClubPage';
 import ChoiceEventPage from '@/pages/event/ChoiceEventPage/ChoiceEventPage';
 import EventDetailPage from '@/pages/event/EventDetailPage/EventDetailPage';
 import SubmitFormPage from '@/pages/event/SubmitFormPage/SubmitFormPage';
@@ -77,7 +77,7 @@ const router = createBrowserRouter([
               },
               {
                 path: 'manage',
-                element: <ManageClubPage />,
+                element: <ClubManagePage />,
               },
               {
                 path: '',
